members/rgd/data-sources/rgd-api/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.104.1
graphql-core==3.2.6
greenlet==3.2.4
h11==0.16.0
idna==3.10
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.27.0
strawberry-graphql==0.200.0
typing_extensions==4.14.1
uvicorn==0.24.0
